+++ server/service/src/cold_chain/tests/mod.rs
@@ -1,2 +1,4 @@
 #[cfg(test)]
 mod query_temperature_log;
+#[cfg(test)]
+mod query_temperature_breach;
