+++ server/service/src/cold_chain/tests/query_temperature_breach.rs
@@ -11,15 +11,12 @@ mod query {
     use crate::{service_provider::ServiceProvider, ListError, SingleRecordError};
 
     #[actix_rt::test]
-    async fn temperature_breach_service_pagination() {
-        let (_, connection, connection_manager, _) = setup_all(
-            "test_temperature_breach_service_pagination",
-            MockDataInserts::all(),
-        )
-        .await;
+    async fn cold_chain_service_pagination() {
+        let (_, connection, connection_manager, _) =
+            setup_all("test_cold_chain_service_pagination", MockDataInserts::all()).await;
 
         let service_provider = ServiceProvider::new(connection_manager, "app_data");
-        let service = service_provider.temperature_breach_service;
+        let service = service_provider.cold_chain_service;
 
         assert_eq!(
             service.temperature_breaches(
@@ -49,7 +46,7 @@ mod query {
     }
 
     #[actix_rt::test]
-    async fn temperature_breach_service_single_record() {
+    async fn cold_chain_service_single_record() {
         let (_, _, connection_manager, _) = setup_all(
             "test_temperature_breach_single_record",
             MockDataInserts::all(),
@@ -58,7 +55,7 @@ mod query {
 
         let service_provider = ServiceProvider::new(connection_manager, "app_data");
         let context = service_provider.basic_context().unwrap();
-        let service = service_provider.temperature_breach_service;
+        let service = service_provider.cold_chain_service;
 
         assert_eq!(
             service.get_temperature_breach(&context, "invalid_id".to_owned()),
@@ -77,12 +74,12 @@ mod query {
     }
 
     #[actix_rt::test]
-    async fn temperature_breach_service_filter() {
+    async fn cold_chain_service_filter() {
         let (_, connection, connection_manager, _) =
             setup_all("test_temperature_breach_filter", MockDataInserts::all()).await;
 
         let service_provider = ServiceProvider::new(connection_manager, "app_data");
-        let service = service_provider.temperature_breach_service;
+        let service = service_provider.cold_chain_service;
 
         let result = service
             .temperature_breaches(
@@ -128,12 +125,12 @@ mod query {
     }
 
     #[actix_rt::test]
-    async fn temperature_breach_service_sort() {
+    async fn cold_chain_service_sort() {
         let (mock_data, connection, connection_manager, _) =
             setup_all("test_temperature_breach_sort", MockDataInserts::all()).await;
 
         let service_provider = ServiceProvider::new(connection_manager, "app_data");
-        let service = service_provider.temperature_breach_service;
+        let service = service_provider.cold_chain_service;
         // Test StartDatetime sort with default sort order
         let result = service
             .temperature_breaches(
