+++ server/service/src/cold_chain/validate.rs
@@ -1,6 +1,7 @@
 use repository::{
     EqualFilter, RepositoryError, StorageConnection, TemperatureLogFilter,
-    TemperatureLogRepository, TemperatureLogRow, TemperatureLogRowRepository,
+    TemperatureLogRepository, TemperatureLogRow, TemperatureLogRowRepository, TemperatureBreachFilter,
+    TemperatureBreachRepository, TemperatureBreachRow, TemperatureBreachRowRepository,
 };
 
 pub fn check_temperature_log_does_not_exist(
@@ -18,4 +19,22 @@ pub fn check_temperature_log_exists(
     connection: &StorageConnection,
 ) -> Result<Option<TemperatureLogRow>, RepositoryError> {
     Ok(TemperatureLogRowRepository::new(connection).find_one_by_id(id)?)
+}
+
+
+pub fn check_temperature_breach_does_not_exist(
+    id: &str,
+    connection: &StorageConnection,
+) -> Result<bool, RepositoryError> {
+    let temperature_breaches = TemperatureBreachRepository::new(connection)
+        .query_by_filter(TemperatureBreachFilter::new().id(EqualFilter::equal_to(id)))?;
+
+    Ok(temperature_breaches.len() == 0)
+}
+
+pub fn check_temperature_breach_exists(
+    id: &str,
+    connection: &StorageConnection,
+) -> Result<Option<TemperatureBreachRow>, RepositoryError> {
+    Ok(TemperatureBreachRowRepository::new(connection).find_one_by_id(id)?)
 }
