--- server/service/src/temperature_breach/tests/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-mod query;
